[arch] separate addon into modules
