+++ mcdata.py
@@ -1,0 +1,20 @@
+
+# Arrays for ENUM properties
+# Array to store different section type
+mc_section_type_list = [
+                ("DEFAULT","Standard","A simple collection of properties that can be added right clicking on fields -> Add Property to the Menu"),
+                ("COLLECTION","Collection List","Right clicking on them in the Outliner, you can add collections whose elements can be shown/hidden in the Menu. Only one collection will be shown at the same time.\nIdeal for: Outfit lists","OUTLINER_COLLECTION",1)
+            ]
+# Array to store possible icons to be used by properties and sections
+mc_icon_list = [
+                ("NONE","No Icon","No Icon"),
+                ("USER", "Face", "Face","USER",1),
+                ("HAIR", "Hair", "Hair","HAIR",2),
+                ("MOD_CLOTH", "Cloth", "Cloth","MOD_CLOTH",3),
+                ("MATERIAL", "Material", "Material","MATERIAL",4),
+                ("ARMATURE_DATA", "Armature", "Armature","ARMATURE_DATA",5),
+                ("MOD_ARMATURE", "Armature", "Armature","MOD_ARMATURE",6),
+                ("EXPERIMENTAL", "Experimental", "Experimental","EXPERIMENTAL",7),
+                ("WORLD", "World", "World","WORLD",8),
+                ("PARTICLEMODE", "Comb", "Comb","PARTICLEMODE",9)
+            ]
